[Refactoring]Remove init.rs
Fixes: #101
