--- bootx64/src/init.rs
@@ -1,22 +1,0 @@
-use uefi::prelude::Boot;
-use uefi::prelude::SystemTable;
-use uefi::ResultExt;
-
-pub fn uefi(system_table: &SystemTable<Boot>) -> () {
-    initialize_uefi_utilities(&system_table);
-    reset_console(&system_table);
-    info!("Hello World!");
-}
-
-/// Initialize uefi-rs services. This includes initialization of GlobalAlloc, which enables us to
-/// use Collections defined in alloc module, such as Vec and LinkedList.
-fn initialize_uefi_utilities(system_table: &SystemTable<Boot>) -> () {
-    uefi_services::init(system_table).expect_success("Failed to initialize_uefi_utilities");
-}
-
-fn reset_console(system_table: &SystemTable<Boot>) -> () {
-    system_table
-        .stdout()
-        .reset(false)
-        .expect_success("Failed to reset stdout");
-}
